refactor(editor): update YouTubeUploader component to YouTubeUploaderV2
