+++ apps/frontend/apps/lyra-app/components/editor/extensions/embed/youtube/YouTubeUploader.ts
@@ -2,7 +2,7 @@ import { type Component } from 'vue';
 import { mergeAttributes, Node, type NodeConfig, type NodeViewProps } from '@tiptap/core';
 import { VueNodeViewRenderer } from '@tiptap/vue-3';
 
-import YouTubeUploader from './YouTubeUploaderV2.vue';
+import YouTubeUploader from './YouTubeUploader.vue';
 
 export default Node.create({
   name: 'youtube-uploader',
